Looked through JS and started changing titles and colors and removing sections
Did some more styling work also looking through the javascript to see what nepal did. I am making this website my own so I am trying to modify some sections and remove others so a good chunk of the HTML and JS is going to not be needed. Im on the fence on removing it or not as keeping it may be a good reference for creating new pages in the future.